+++ docs/src/app/pages/docs/slides/app-slides-concept/app-slides-concept.tsx
@@ -21,6 +21,21 @@ export class AppSlidesConcept {
     }
   }
 
+  playPauseVideo(): Promise<void> {
+    return new Promise<void>(async (resolve) => {
+      const element: any = this.el.querySelector('deckgo-slide-video');
+
+      if (!element) {
+        resolve();
+        return;
+      }
+
+      await element.toggle();
+
+      resolve();
+    })
+  }
+
   render() {
     return [
       <app-navigation></app-navigation>,
@@ -98,19 +113,33 @@ export class AppSlidesConcept {
 </div>
 
 <ul>
+<li>Slide: <a href="/slides/bigimg">Big Image</a></li>
+</ul>
+<div class="container ion-margin">
+  <deckgo-deck embedded={true}>
+    <deckgo-slide-big-img
+             img-src="https://raw.githubusercontent.com/deckgo/deckdeckgo/master/webcomponents/slides/big-img/showcase/big-deckdeckgo-h.jpg"
+             img-divisions="900;1500;2200"
+             axis="x"
+             reverse>
+    </deckgo-slide-big-img>
+  </deckgo-deck>
+</div>
+
+<ul>
 <li>Slide: <a href="/slides/chart">Chart</a></li>
 </ul>
 <div class="container ion-margin">
   <deckgo-deck embedded={true}>
-    <deckgo-slide-chart width={200} height={100} src="https://raw.githubusercontent.com/fluster/deckdeckgo-charts/master/showcase/data-pie-chart.csv">
+    <deckgo-slide-chart width={200} height={100} src="https://raw.githubusercontent.com/deckgo/deckdeckgo/master/webcomponents/charts/showcase/data-pie-chart.csv">
       <h1 slot="title">slot="title"</h1>
     </deckgo-slide-chart>
     <deckgo-slide-chart width={200} height={100} type="line" y-axis-domain="extent" date-pattern="dd.MM.yyyy"
-                        src="https://raw.githubusercontent.com/fluster/deckdeckgo-charts/master/showcase/data-line-chart-to-compare.csv">
+                        src="https://raw.githubusercontent.com/deckgo/deckdeckgo/master/webcomponents/charts/showcase/data-line-chart-to-compare.csv">
       <h1 slot="title">slot="title"</h1>
     </deckgo-slide-chart>
     <deckgo-slide-chart width={200} height={100}
-                        type="bar" src="https://raw.githubusercontent.com/fluster/deckdeckgo-charts/master/showcase/data-bar-chart-to-compare.csv"
+                        type="bar" src="https://raw.githubusercontent.com/deckgo/deckdeckgo/master/webcomponents/charts/showcase/data-bar-chart-to-compare.csv"
                         style={{'--deckgo-chart-fill-color-bar1': 'var(--ion-color-primary)', '--deckgo-chart-fill-color-bar2': 'var(--ion-color-secondary)', '--deckgo-chart-fill-color-bar3': 'var(--ion-color-tertiary)'}}
                         >
       <h1 slot="title">slot="title"</h1>
@@ -126,6 +155,18 @@ export class AppSlidesConcept {
     <deckgo-slide-youtube src="https://www.youtube.com/watch?v=oUOjJIfPIjw">
       <h1 slot="title">slot="title"</h1>
     </deckgo-slide-youtube>
+  </deckgo-deck>
+</div>
+
+<ul>
+<li>Slide: <a href="/slides/video">Video</a></li>
+</ul>
+<div class="container ion-margin">
+  <deckgo-deck embedded={true}>
+    <deckgo-slide-video src="https://media.giphy.com/media/vv41HlvfogHAY/giphy.mp4">
+      <h1 slot="title">A Gif as video</h1>
+      <button slot="actions" onClick={() => this.playPauseVideo()}>Play/pause</button>
+    </deckgo-slide-video>
   </deckgo-deck>
 </div>