+++ webcomponents/slides/utils/src/index.ts
@@ -1,4 +1,5 @@
 export {DeckdeckgoSlide} from './interfaces/deckdeckgo-slide';
+export {DeckdeckgoSlidePlay} from './interfaces/deckdeckgo-slide-play';
 export {DeckdeckgoComponent} from './interfaces/deckdeckgo-component';
 
 export * from './utils/deckdeckgo-component-utils';